Segunda Entrega

+++ src/services/adminService.ts
@@ -2,6 +2,13 @@ import {AdminRepository} from "../repository/adminRepository";
 import * as bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
 import adminController from "../controller/adminController";
+import {ShieldAlreadyExistsError} from "../errors/ShieldAlreadyExistsError";
+import {Admin} from "../models/Admin";
+import {EmailAlreadyInUseError} from "../errors/ EmailAlreadyInUseError";
+import {RequiredFieldsAreMissingError} from "../errors/RequiredFieldsAreMissingError";
+import {Shield} from "../models/Shield";
+import {ShieldNotFoundError} from "../errors/ShieldNotFoundError";
+import {AdminNotFoundError} from "../errors/AdminNotFoundError";
 
 
 
@@ -16,18 +23,33 @@ export class AdminService {
     }
 
 
-    async createAdmin(name: string, email: string, password: string) {
-        return await this.adminRepository.createAdmin(name, email, password);
-    }
+    // async createAdmin(name: string, email: string, password: string) {
+    //     return await this.adminRepository.createAdmin(name, email, password);
+    // }
 
-    async createAdminCrypt(name: string, email: string, password: string) {
+    async createAdminCrypt(data: Partial<Admin>) {
 
-        const hashedPassword = await bcrypt.hash(password, 10);
 
-        return await this.adminRepository.createAdmin(name, email, hashedPassword);
+        const { name, email, password } = data;
+
+        const existing = await this.adminRepository.getAdminByEmail(<string>data.email);
+        if (existing) {
+            throw new EmailAlreadyInUseError();
+        }
+
+        if (hasEmptyRequiredFields(data)) {
+            throw new RequiredFieldsAreMissingError();
+        }
+        const hashedPassword = await bcrypt.hash(<string>data.password, 10);
+
+        return await this.adminRepository.createAdmin(data);
     }
 
     async getAdminById(id: string) {
+        const armor = await this.adminRepository.getAdminById(id);
+        if (!armor) {
+            throw new AdminNotFoundError();
+        }
         return await this.adminRepository.getAdminById(id);
     }
 
@@ -40,10 +62,18 @@ export class AdminService {
     }
 
     async updateAdmin(id: string, data: Partial<{ name: string; email: string; password: string }>) {
+        const admin = await this.adminRepository.getAdminById(id);
+        if (!admin) {
+            throw new AdminNotFoundError();
+        }
         return await this.adminRepository.updateAdmin(id, data);
     }
 
     async deleteAdmin(id: string) {
+        const success = await this.adminRepository.deleteAdmin(id);
+        if (!success) {
+            throw new AdminNotFoundError();
+        }
         return await this.adminRepository.deleteAdmin(id);
     }
 
@@ -62,5 +92,16 @@ export class AdminService {
 
         return { user, token };
     }
+
+
+}
+
+function hasEmptyRequiredFields(data: Partial<Admin>): boolean {
+
+    const requiredKeys: (keyof Admin)[] = ["name", "email", "password"];
+    return requiredKeys.some(key => {
+        const value = data[key];
+        return value === undefined || value === null || value === "";
+    });
 }
